requirements.txt
asttokens==2.0.5
backcall==0.2.0
brotlipy==0.7.0
certifi==2022.9.24
cffi==1.15.0
charset-normalizer==2.1.1
cryptography==37.0.2
cycler==0.11.0
decorator==5.1.1
executing==0.8.3
filelock==3.8.0
fonttools==4.38.0
huggingface-hub==0.11.0
idna==3.4
ipython==8.6.0
jedi==0.18.1
joblib==1.2.0
kaggle==1.5.12
kiwisolver==1.4.4
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-service==2.4.0
olefile==0.46
packaging==21.3
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pip==22.2.2
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
setuptools==65.5.0
six==1.16.0
stack-data==0.2.0
text-unidecode==1.3
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.1
traitlets==5.1.1
transformers==4.24.0
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
wheel==0.37.1
